Keep track of changes in Activity description
The user used to be allowed to dismiss the Activity Detail screen
without confirmation, even if there were unsaved changes to the
activity's description. This commit ensures that changes to an
activity's description are taken into account when deciding whether to
allow a user-initiated pull-down of the view. In simpler words, if
activity description has changed, ask the user for confirmation before
dismissing.

Additionally, ensure the "Add" button is enabled only when there are
unsaved changes to the activity name. Changes in activity description
have no say on the status of the button.