Rename property of Activity
Since Activity used to conform to 'CustomStringConvertible', and had to
provide the 'description' property, we were not able to use
'description' as the property name for the activity's description. We
were forced to use 'activityDescription' instead.

However, it turns out that, for debugging purposes, it's better to
conform to 'CustomDebugStringConvertible' (and provide the
'debugDescription' property instead.) This frees us up to use
'description' as the property name for the activity's description.