Improve input validation
When doing input validation in the Activity Detail screen, replace the
use of a generic `ErrorMessage` with a more specific
`ActivityCreationError`, which conforms to `LocalizedError` in order to
provide human-readable explanations of what went wrong.

This greatly simplifies the error handling code by moving most of the
implementation details to 'ActivityCreationError.swift'.