Stay on editing mode until saving is successful
On the Activity Detail screen, do not exit editing mode as soon as the
user taps on the "Done" button item. Instead, rely on the fact that the
whole modal screen is dismissed when the save operation is successful.
So there is no actual need to ever *leave* editing mode.

Before this commit, the "Done" button would immediately turn into "Edit"
while the save operation was still unfinished; i.e. while the
confirmation alert was on-screen, or while the input validation alert
was on-screen.