Refactor dismissal of Activity Detail screen (w/o saving new activity)
For the "new activity" flow, when the user cancels the creation of a new
activity without saving changes:

Rather than letting the Activity Detail view controller perform its own
dismissal itself, we'll move that responsibility to the presenting view
controller, i.e. the Landing view controller. This makes for cleaner
code and better separation of concerns.

As a side-effect, activity deletion is now performed in the Landing view
controller.