Refactor dismissal of Activity Detail screen (saving new activity)
For the "new activity" flow, when the user confirms the creation of a
new activity:

Rather than letting the Activity Detail view controller perform its own
dismissal itself, we'll move that responsibility to the presenting view
controller, i.e. the Landing view controller. This makes for cleaner
code and better separation of concerns.

The actual saving of changes is still done in the Activity Detail view
controller, since it involves working with private UI properties.