Make "Done" switch editable only in Editing mode
The "Done" switch used to be enabled all the time, making it possible to
change the status of an activity without having to enter Editing mode
first. This led to non-intuitive behavior because the user could tap on
'Cancel', yet the status change would still be applied.

By making this code change, the user is now required to enter Editing
mode before they can modify the activity status. This makes for a more
consistent user experience, e.x. if the 'Cancel' button is tapped, any
changes are lost.