Fix bug in the activity saving algorithm
The saving algorithm used to simply append the received activity to the
`allActivities` array, without checking if the activity was already in
the array to being with.

This commit changes that behavior, ensuring that we first check if the
activity already exists, in which case we update it. Only if it doesn't
exist do we append it to the end of the array.