Update the Landing screen as soon as changes occur
Prior to this commit, there was a noticeable delay in updating the
Landing screen after the user returned from Activity Detail screen after
having made modifications to the list of activities.

Now, the Landing screen will be updated as soon as any changes to
activities are saved. This way, when the user comes back to the list of
activities, everything will be up-to-date.