Fix abstraction in Landing and ActivityDetail view controllers
Prior to this commit, the view controllers had a poor implementation of
the "abstraction" principle. Both the `LandingViewController` and the
`ActivitiesViewController` knew about the existence of
`ActivityDetailViewController`, which means they both had a dependency
on it. This breaks the principle that says: "a thing should only really
know about objects that it contains, and then only about one layer of
abstraction down."

`LandingViewController` and `ActivitiesViewController` were
unnecessarily aware of the context in which they were used. This
violates the whole "don't know your context" principle.

This improvement consists on extracting the concern of "sequence" out of
the aforementioned view controllers, and into a new container view
controller named `MainViewController`. In general, a view controller
should manage either sequence or UI, but not both.