Use delegation instead of closure properties
Prior to this commit, `ActivityDetailViewController` and
`ActivitiesViewController` relied on closure properties (named 'onSave'
and 'onDelete', respectively) in order to update the list of activities
after an activity was saved/deleted.

From now on, the aforementioned view controllers will instead rely on
delegate protocols to notify to their delegate that an activity has been
saved/deleted. The delegate will have the responsibility to update the
list of activities.