Split `FileActivityDataStore` into two caretakers
By splitting `FileActivityDataStore` into two separate caretakers,
`FileActivityDataStore` itself and `DiskCaretaker`, we considerably
reduce the size of the original class.

The reasoning behind this change is rooted in the application of the
Memento design pattern. `DiskCaretaker` acts as a generic caretaker that
is able to store originators (objects that conform to Codable) into
persistent storage in the filesystem; whereas `FileActivityDataStore` is
a more specific caretaker tailored for managing activities only.