Refactor `ActivityDetail` screen to use the Builder pattern
We are extracting all the activity-building logic out of the
`ActivityDetailViewController`, and into a new class called
`ActivityBuilder`. The builder takes care of input validation, as well
as keeping track of any changes that are made to its initial state.

An `ActivityBuilder` can be instantiated in one of two manners:

* Without an activity, in which case the builder's properties will
  default to values appropriate for the creation of new activities.
* With an activity, in which case the builder's properties will be
  initialized with the data from the received activity. This allows us
  to use the builder for the modification of existing activities (i.e.
  you instantiate a builder with an existing activity, use the builder
  to modify the activity data as needed, and finally call `build()` to
  get a replacement activity.)