Rewrite observation of `activitiesCount` using KVO
We're replacing the previous implementation of property observation,
which relied on a custom `Observable` wrapper, with Key-Value Observing
(KVO). The reason for this change is that the wrapper does not match all
the functionality offered by KVO, which we're going to need in order to
implement property observation for the `activities` property.

As a necessary side effect of this change, we've had to expose several
of our classes/protocols to the Objective-C runtime, since this is a
prerequisite for KVO.