Refactor property observation of `activitiesCount`
We're rewriting our implementation of the observation of the
`activitiesCount` property so that the
`observe(_:options:changeHandler:)` method is called on the actual
subject that owns the observable property.

Before this change, `observe(_:options:changeHandler:)` was being called
on the observer itself, and thus the first parameter to the
`changeHandler` callback was unusable (because it was an instance of the
observer rather than the subject). Now that we're properly observing the
subject, we can access it inside the callback to extract its latest
state.

For this to work, we had to wrap each call to
`observe(_:options:changeHandler:)` inside a generic wrapper method
which constrains the subject to be `NSObject & <some_protocol>`.