Make `Activity` the identity object of `ActivityDetailViewController`
`ActivityDetailViewController`'s identity model object should be
`Activity` because it is an activity what gives this view controller its
identity (letting the view controller understand its position in the
program and how to talk to the model layer.)

`ActivityDetailViewController` should be responsible for the creation of
new instances of `Activity`, so we're taking that responsibility away
from the `ActivityBuilder` and moving it into the view controller.

`ActivityDetailViewController` gains access to its `Activity` identity
object via initializer injection.