Add child view controller at `viewDidLoad()`
Adding `ActivitiesViewController` as a child of `LandingViewController`
at `viewWillLayoutSubviews()` makes it harder to properly implement
state restoration later (it is too late). The ideal place to do it is
inside `viewDidLoad()`.

Since during `viewDidLoad()`, the view controller's view has yet to be
added to the view hierarchy, we are required to remove the precondition
validation that ensured that `containingView` was already part of the
parent view controller's view hierarchy.