Remove unnecessarily duplicate code
There was a lot of unnecessarily duplicated code in both
`ActivityDataSource` and `ActivityRepository`. There was not a clear
separation between the responsibilities of each. Hence, we're removing
all that duplication.

The `ActivityRepository` class will now be in charge of in-memory
activity management, whereas the `ActivityDataSource` will take care of
preserving and reading those activities to/from some kind of persistent
storage (a document in the filesystem, a database, or other).